refactor: split out `inferEntitySummariesFromWebPage` method
